+++ src/game/enemy-factory.ts
@@ -1,34 +1,23 @@
 /* Converted from TimePilot.EnemyFactory.js (AMD) to ESM TypeScript. */
 import CONSTS from "./constants";
 import Enemy from "./enemy";
-import SoundEngine from "./engine/Sound";
 import type {
-  BulletFactoryInstance,
   EnemyConfig,
   EnemyData,
   EnemyFactoryInstance,
   EnemyInstance,
   GameDataStore,
   Heading,
-  PlayerInstance,
 } from "./types";
 
 class EnemyFactory implements EnemyFactoryInstance {
-  private _bullets: BulletFactoryInstance;
   private _context: GameDataStore;
   private _enemies: EnemyInstance[] = [];
-  private _explosionSound: SoundEngine;
   private _level: number;
-  private _player: PlayerInstance;
 
   constructor(context: GameDataStore) {
     this._context = context;
     this._level = context._level;
-    this._player = context._player;
-    this._bullets = context._bullets;
-    this._explosionSound = new SoundEngine(
-      this.getLevelData().explosion.sound.src
-    );
   }
 
   create(posX: number, posY: number, heading: Heading): void {
@@ -59,42 +48,8 @@ class EnemyFactory implements EnemyFactoryInstance {
     return this._enemies.map((enemy) => enemy.getData() as EnemyData);
   }
 
-  detectCollision(): void {
-    const bullets = this._bullets.getData();
-    const playerData = this._player.getData();
-
-    this._enemies.forEach((enemy) => {
-      if (!enemy.isAlive || !playerData.isAlive) {
-        return;
-      }
-
-      if (
-        enemy.detectCollision(
-          playerData.posX,
-          playerData.posY,
-          CONSTS.player.hitRadius
-        )
-      ) {
-        enemy.kill();
-        this._explosionSound.stop();
-        this._explosionSound.play();
-        this._player.kill();
-      }
-
-      bullets.forEach((bullet) => {
-        if (
-          enemy.detectCollision(
-            bullet.posX + this._player.getData().posX,
-            bullet.posY + this._player.getData().posY,
-            CONSTS.player.projectile.size
-          )
-        ) {
-          enemy.kill();
-          this._explosionSound.stop();
-          this._explosionSound.play();
-        }
-      });
-    });
+  getEntities(): EnemyInstance[] {
+    return [...this._enemies];
   }
 
   cleanup(): void {